+++ R/ColumnLinkedMatrix.R
@@ -40,7 +40,7 @@ subset.ColumnLinkedMatrix <- function(x, i, j, ..., drop) {
         sortedColumns <- j
     }
     # Compute node inventory
-    globalIndex <- index(x)[sortedColumns, , drop = FALSE]
+    globalIndex <- index(x, sortedColumns)
     whatChunks <- unique(globalIndex[, 1])
     # If there are several chunks involved, aggregate the result in a separate
     # matrix, otherwise pass through result
@@ -219,34 +219,41 @@ rbind.ColumnLinkedMatrix <- function(..., deparse.level = 1) {
 #' @export
 nodes.ColumnLinkedMatrix <- function(x) {
     n <- nNodes(x)
-    OUT <- matrix(nrow = n, ncol = 3, NA)
-    colnames(OUT) <- c("node", "col.ini", "col.end")
+    OUT <- matrix(integer(), nrow = n, ncol = 3, dimnames = list(NULL, c("node", "col.ini", "col.end")))
     end <- 0
-    for (i in 1:n) {
+    for (node in 1:n) {
         ini <- end + 1
-        end <- ini + ncol(x[[i]]) - 1
-        OUT[i, ] <- c(i, ini, end)
+        end <- ini + ncol(x[[node]]) - 1
+        OUT[node, ] <- c(node, ini, end)
     }
     return(OUT)
 }
 
 
 #' @export
-index.ColumnLinkedMatrix <- function(x) {
+index.ColumnLinkedMatrix <- function(x, j = NULL) {
     nodes <- nodes(x)
-    nColIndex <- nodes[nrow(nodes), 3]
-    INDEX <- matrix(nrow = nColIndex, ncol = 3)
-    colnames(INDEX) <- c("node", "col.global", "col.local")
-    INDEX[, 2] <- 1:nColIndex
-    end <- 0
-    for (i in 1:nrow(nodes)) {
-        nColChunk <- nodes[i, 3] - nodes[i, 2] + 1
-        ini <- end + 1
-        end <- ini + nColChunk - 1
-        INDEX[ini:end, 1] <- i
-        INDEX[ini:end, 3] <- 1:nColChunk
+    if (is.null(j)) {
+        p <- nodes[nrow(nodes), 3]
+        OUT <- matrix(nrow = p, ncol = 3, dimnames = list(NULL, c("node", "col.global", "col.local")))
+        OUT[, 2] <- seq_len(ncol(x))
+        end <- 0
+        for (node in 1:nrow(nodes)) {
+            nodeSize <- nodes[node, 3] - nodes[node, 2] + 1
+            ini <- end + 1
+            end <- ini + nodeSize - 1
+            OUT[ini:end, 1] <- node
+            OUT[ini:end, 3] <- seq_len(nodeSize)
+        }
+    } else {
+        OUT <- t(sapply(j, function(globalIndex) {
+            node <- which(globalIndex >= nodes[, 2] & globalIndex <= nodes[, 3])
+            localIndex <- globalIndex - nodes[node, 2] + 1
+            c(node, globalIndex, localIndex)
+        }))
+        dimnames(OUT) <- list(NULL, c("node", "row.global", "row.local"))
     }
-    return(INDEX)
+    return(OUT)
 }