+++ R/LinkedMatrix.R
@@ -103,11 +103,16 @@ nodes <- function(x) {
 #' Maps each column or row index of a linked matrix to the column or row index
 #' of its corresponding node.
 #'
+#' If \code{j} for \code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}}
+#' or \code{i} for \code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} is
+#' passed, it will only generate entries for the given indices.
+#'
 #' @param x Either a \code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}}
 #'   or a \code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} object.
+#' @param ... Additional arguments (see Details).
 #' @return A matrix.
 #' @export
-index <- function(x) {
+index <- function(x, ...) {
     UseMethod("index")
 }